backend/requirements.txt
starlette
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
asyncpg==0.30.0

# Redis
redis==5.2.0
aioredis==2.0.1

# AI/ML
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
openai==1.59.6
sentence-transformers==3.3.1
qdrant-client==1.12.1

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
python-magic==0.4.27

# Utilities
uuid7
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Development
black==24.10.0
ruff==0.8.4
mypy==1.13.0